fix hero training time

+++ install/data/sql.sql
@@ -783,7 +783,7 @@ CREATE TABLE IF NOT EXISTS `%PREFIX%hero` (
   `defencebonus` tinyint(1) unsigned NOT NULL,
   `regeneration` tinyint(1) unsigned NOT NULL,
   `autoregen` int(2) NOT NULL,
-  `trainingtime` mediumint(3) unsigned NOT NULL,
+  `trainingtime` int(11) unsigned NOT NULL,
   PRIMARY KEY (`heroid`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 AUTO_INCREMENT=1;
